Updated box and URL. Made install-sh smarter.

+++ install-trove.sh
@@ -1,48 +1,59 @@
 #!/bin/bash
 
+echo Updating apt and installing prerequisites.
 export DEBIAN_FRONTEND=noninteractive
-apt-get update -y
-apt-get install -y --force-yes git-core
+apt-get update -y -q
+apt-get install -y -q git-core libxml2-dev libxslt1-dev
 
-# Code should be shared at /trove.
-# If it isn't, go get it.
-# Get the code. If /trove isn't populated, go get the pieces.
-mkdir -p /opt/stack
+
+echo Linking Trove codebase.
 SHARE=/trove
-mkdir -p $SHARE
+mkdir -p /opt/stack $SHARE
 pushd $SHARE
     for REPO in python-troveclient trove trove-integration; do
         if [ ! -d "$SHARE/$REPO" ]; then
             git clone git://github.com/openstack/$REPO.git
         fi
-        ln -s $SHARE/$REPO /opt/stack/$REPO
+        if [ ! -e "/opt/stack/$REPO" ]; then
+            ln -s $SHARE/$REPO /opt/stack/$REPO
+        fi
     done
 popd
 
-# create user trove
-useradd trove -d /home/trove -s /bin/bash -m
-echo trove:trove | chpasswd
-sed -i '/^%sudo/a\trove ALL=(ALL) NOPASSWD:ALL' /etc/sudoers
-ln -s /opt/stack/trove-integration /home/trove/trove-integration
-sed -i '$a\cd /home/trove/trove-integration/scripts' /home/trove/.bashrc
 
-chown trove /opt/stack
+echo Creating user trove.
+if [ ! -d /home/trove ]; then
+    useradd trove -d /home/trove -s /bin/bash -m
+    echo trove:trove | chpasswd
+    sed -i '/^%sudo/a\trove ALL=(ALL) NOPASSWD:ALL' /etc/sudoers
+    ln -s /opt/stack/trove-integration /home/trove/trove-integration
+    sed -i '$a\cd /home/trove/trove-integration/scripts' /home/trove/.bashrc
+    chown trove /opt/stack
+fi
 
-# Install Django manually, as pip is acting a little slow
-pushd /tmp
-    wget http://pypi.python.org/packages/source/D/Django/Django-1.5.1.tar.gz#md5=7465f6383264ba167a9a031d6b058bff -O Django.tgz
-    tar xzf Django.tgz
-    pushd Django-1.5.1
-        python setup.py install
+
+echo Installing Django manually.
+if [ ! -e "/usr/local/lib/python2.7/dist-packages/django" ]; then
+    pushd /tmp
+        wget http://pypi.python.org/packages/source/D/Django/Django-1.5.1.tar.gz#md5=7465f6383264ba167a9a031d6b058bff -O Django.tgz -q
+        tar xzf Django.tgz
+        pushd Django-1.5.1
+            python setup.py install > /dev/null 2>&1
+        popd
+        rm -rf Django*
     popd
-    rm -rf Django*
-popd
+fi
 
-# Create fix-iptables.sh
+
+echo Creating fix-iptables.sh
 FIXSH="fix-iptables.sh"
-pushd /opt/stack/trove-integration/scripts
-    echo "#!/bin/bash" > $FIXSH
-    echo "sudo iptables -t nat -A POSTROUTING -s 10.0.0.0/24 -o eth0 -j MASQUERADE" >> $FIXSH
-    chmod +x $FIXSH
+if [ ! -e "$FIXSH" ]; then
+    pushd /opt/stack/trove-integration/scripts
+        echo "#!/bin/bash" > $FIXSH
+        echo "sudo iptables -t nat -A POSTROUTING -s 10.0.0.0/24 -o eth0 -j MASQUERADE" >> $FIXSH
+        chmod +x $FIXSH
+    popd
+fi
+
 
 echo Installed.
